organising/adding mvvm on notifications view
